--- src/main/java/com/example/cp_main_be/avatar/dto/request/AvatarRequest.java
@@ -1,3 +1,0 @@
-package com.example.cp_main_be.avatar.dto.request;
-
-public class AvatarRequest {}
