+++ src/main/java/com/example/cp_main_be/domain/avatar/domain/Avatar.java
@@ -1,0 +1,21 @@
+package com.example.cp_main_be.domain.avatar.domain;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+@Entity
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class Avatar {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  @Column(name = "avatar_id")
+  private Long id;
+
+  @Column(name = "image_url")
+  private String imageUrl;
+}
