+++ src/main/java/com/example/cp_main_be/domain/avatar/domain/repository/AvatarRepository.java
@@ -1,0 +1,6 @@
+package com.example.cp_main_be.domain.avatar.domain.repository;
+
+import com.example.cp_main_be.domain.avatar.domain.Avatar;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface AvatarRepository extends JpaRepository<Avatar, Long> {}
