+++ src/main/java/com/example/cp_main_be/domain/avatar/dto/request/AvatarRequest.java
@@ -1,0 +1,3 @@
+package com.example.cp_main_be.domain.avatar.dto.request;
+
+public class AvatarRequest {}
