+++ src/main/java/com/example/cp_main_be/domain/user/domain/UserStatus.java
@@ -1,0 +1,7 @@
+package com.example.cp_main_be.domain.user.domain;
+
+public enum UserStatus {
+  ACTIVE,
+  INACTIVE,
+  BANNED
+}
