--- src/main/java/com/example/cp_main_be/user/domain/UserStatus.java
@@ -1,7 +1,0 @@
-package com.example.cp_main_be.user.domain;
-
-public enum UserStatus {
-  ACTIVE,
-  INACTIVE,
-  BANNED
-}
