--- src/main/java/com/example/cp_main_be/user/domain/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.example.cp_main_be.user.domain.repository;
-
-import com.example.cp_main_be.user.domain.User;
-import java.util.Optional;
-import java.util.UUID;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByUsername(String username);
-
-  Optional<User> findByUuid(UUID uuid);
-}
